requirements.txt
flask
requests
pyirsdk
